move upgrade handlers to v8.0
