+++ protocol/testing/version/VERSION_CURRENT
@@ -1,1 +1,1 @@
-v7.1
+v8.0
