+++ protocol/testing/version/VERSION_FULL_NAME_PREUPGRADE
@@ -1,1 +1,1 @@
-v7.0.2
+v6.0.0-dev999
